[bugfix] fixing type cast for test suite

+++ unittests/main_unittest.c
@@ -21,7 +21,7 @@ int main(i32 argc, char** argv)
 	RUN_TEST_SUITE(id_tests);
 
 	const char* testResultMessage = "********** TEST RESULT **********";
-	i32			width			  = strlen(testResultMessage); // TODO: use self-implemented strlen later
+	i32			width			  = (i32)strlen(testResultMessage); // TODO: use self-implemented strlen later
 
 	if (totalTestsCount == passedTestsCount)
 	{
@@ -35,8 +35,8 @@ int main(i32 argc, char** argv)
 		ntt_FormatMessage(
 			resultMessage, sizeof(resultMessage), "Some tests failed! (%llu/%llu)", passedTestsCount, totalTestsCount);
 
-		i32 leftPadding	 = (width - 2 - strlen(resultMessage)) / 2; // TODO: use self-implemented strlen later
-		i32 rightPadding = width - 3 - leftPadding - strlen(resultMessage);
+		i32 leftPadding	 = (width - 2 - (i32)strlen(resultMessage)) / 2; // TODO: use self-implemented strlen later
+		i32 rightPadding = width - 3 - leftPadding - (i32)strlen(resultMessage); // TODO: use self-implemented strlen later
 
 		ntt_ConsoleSetColor(NTT_COLOR_RED);
 		ntt_ConsolePrint("%s\n", testResultMessage);
